+++ data/data_util.py
@@ -6,7 +6,6 @@ from PIL import Image
 import torch
 import torchvision.transforms as transforms
 
-import util
 import gc
 
 def load_hdf5_data(filename, num_classes):
