+++ utils/util.py
@@ -2,6 +2,8 @@
 import os, struct, math
 import numpy as np
 import torch
+import torch.nn as nn
+import torch.nn.functional as F
 
 # util for saving tensors, for debug purposes
 def write_array_to_file(tensor, filename):
@@ -103,3 +105,17 @@ class BNMomentumScheduler(object):
 
         self.last_epoch = epoch
         self.model.apply(self.setter(self.lmbd(epoch)))
+
+
+class WeightedCrossEntropyLoss(nn.Module):
+    def __init__(self, ignore_index=-100):
+        super(WeightedCrossEntropyLoss, self).__init__()
+        self.ignore_index = ignore_index
+
+    def forward(self, inputs, targets, weights):
+        assert inputs.size(0) == targets.size(0) == weights.size(0)
+
+        loss = F.cross_entropy(input=inputs, target=targets, reduction="none", ignore_index=self.ignore_index)
+        loss = torch.mean(loss * weights.float())
+
+        return loss
